Refactor both AdminComponents into one
