+++ src/app/js/formats/link/index.js
@@ -1,5 +1,5 @@
 import Component from './Component';
-import AdminComponent from './AdminComponent';
+import AdminComponent from '../DefaultAdminComponentWithLabel';
 import DefaultFormat from '../DefaultFormat';
 
 export default {
